% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHMBSpatialCorr.R
\docType{data}
\name{FIDs1}
\alias{FIDs1}
\title{A vector of containing the IDs of the population units in sample s1}
\format{
An object of class \code{integer} of length 100.
}
\usage{
data(FIDs1)
}
\description{
A vector of containing the IDs of the population units in sample s1
}
\examples{
data(FIDs1)
}
\keyword{datasets}
