% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHMBSpatialCorr.R
\docType{data}
\name{FIDs2}
\alias{FIDs2}
\title{A vector of containing the IDs of the population units in sample s2}
\format{
An object of class \code{integer} of length 200.
}
\usage{
data(FIDs2)
}
\description{
A vector of containing the IDs of the population units in sample s2
}
\examples{
data(FIDs2)
}
\keyword{datasets}
