% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHMBSpatialCorr.R
\name{getCorrelationSumForResidualErrors}
\alias{getCorrelationSumForResidualErrors}
\title{Compute the sum of the correlations amongst the residual error terms}
\usage{
getCorrelationSumForResidualErrors(
  calculator,
  parm,
  subpopulation = NULL,
  nbThreads = 3
)
}
\arguments{
\item{calculator}{the pointer to the Java instance of the SpatialCorrelationCalculator class}

\item{parm}{the correlation parameter (theta) of a exponential spatial correlation structure, i.e. e^{-d/theta}
where d is the distance between two population units}

\item{subpopulation}{a vector of integers (numerics will be automatically coerced to integers) that stands
for the FID (integers) of the units in the subpopulation. A null value means the correlations are calculated
across the entire population.}

\item{nbThreads}{the number of threads to put on the job. An integer between 1 and 8 (by default is set to 3)}
}
\description{
Compute the sum of the correlations amongst the residual error terms
}
