% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHMBSpatialCorr.R
\name{getCorrelationSumsForTheseUnits}
\alias{getCorrelationSumsForTheseUnits}
\title{Calculate the sums of residual error correlations between the population (or a subpopulation)
and the units of the sample.}
\usage{
getCorrelationSumsForTheseUnits(
  calculator,
  parm,
  sample = c(1, 2),
  subpopulation = NULL,
  nbThreads = 3
)
}
\arguments{
\item{calculator}{the pointer to the Java instance of the SpatialCorrelationCalculator class}

\item{parm}{the correlation parameter (theta) of a exponential spatial correlation structure, i.e. e^{-d/theta}
where d is the distance between two population units}

\item{sample}{1 for sample 1 or 2 for sample 2}

\item{subpopulation}{a vector of integers (numerics will be automatically coerced to integers) that stands
for the FID (integers) of the units in the subpopulation. A null value means the correlations are calculated
across the entire population.}

\item{nbThreads}{the number of threads to put on the job. An integer between 1 and 8 (by default is set to 3)}
}
\value{
a matrix of numerics
}
\description{
Calculate the sums of residual error correlations between the population (or a subpopulation)
and the units of the sample.
}
