% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHMB_SpatialCorr.R
\name{initializeCalculator}
\alias{initializeCalculator}
\title{Initialize the calculator}
\usage{
initializeCalculator(calculator, FIDs, UTM_X, UTM_Y, sample1, sample2)
}
\arguments{
\item{calculator}{the pointer to the Java instance of the SpatialCorrelationCalculator class}

\item{FIDs}{a vector containing the ids of the population units. These ids must be integers.}

\item{UTM_X}{a vector of UTM longitude}

\item{UTM_Y}{a vector of UTM latitude}

\item{sample1}{a vector of integers standing for the ids of the units in sample 1}

\item{sample2}{a vector of integers standing for the ids of the units in sample 2}
}
\value{
nothing
}
\description{
Initialize the calculator
}
