% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHMB_SpatialCorr.R
\docType{data}
\name{jarFilename}
\alias{jarFilename}
\title{Functions to calculate the contribution
of the spatial correlations at the population
or sub population level.}
\format{
An object of class \code{character} of length 1.
}
\usage{
jarFilename
}
\description{
Functions to calculate the contribution
of the spatial correlations at the population
or sub population level.
}
\author{
Mathieu Fortin - June 2022
}
\keyword{datasets}
