% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHMBSpatialCorr.R
\docType{data}
\name{popExample}
\alias{popExample}
\title{A data.frame object to provide an example}
\format{
An object of class \code{data.frame} with 70137 rows and 4 columns.
}
\usage{
data(popExample)
}
\description{
A data.frame object to provide an example
}
\examples{
data(popExample)
}
\keyword{datasets}
