% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHMB_SpatialCorr.R
\name{SpatialCorrelationCalculator}
\alias{SpatialCorrelationCalculator}
\title{Constructor of the SpatialCorrelationCalculator}
\usage{
SpatialCorrelationCalculator(rasterResolutionM, scaleFactor)
}
\arguments{
\item{rasterResolutionM}{the raster resolution  (e.g. 30 m for Landsat-8)}

\item{scaleFactor}{a scaling factor for the correlation function (i.e. 0.001 if the function is based on km)}
}
\value{
a pointer to a Java object of the class SpatialCorrelationCalculator
}
\description{
Constructor of the SpatialCorrelationCalculator
}
